pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "math-tutor"
version = "0.1.0"
description = "Fetch PDFs from a Canvas course and process them with the OpenAI Responses API."
readme = "math_tutor/README.md"
requires-python = ">=3.10"
dependencies = [
    "google-genai>=1.0.0",
    "httpx>=0.27.0",
    "openai>=1.40.0",
    "playwright>=1.45.0",
]

[project.scripts]
math-tutor = "math_tutor.cli:main"
math-tutor-build-site = "math_tutor.site_builder:main"
math-tutor-backfill-response-html = "math_tutor.backfill_response_html:main"
math-tutor-build-challenges = "math_tutor.challenge_builder:main"
math-tutor-generate-mcq = "math_tutor.mcq_generator:main"

[tool.setuptools]
packages = ["math_tutor"]
